refactor: overhaul colour pipeline & documentation
- Replace CIE2000 delta-E (CIELabColor) with Oklab for perceptually uniform
  colour matching; add OklabColor, RGBColorDouble utility classes
- Replace monolithic block_data.json with dynamic blockset system:
  online blocksets (All/Default/Grayscale/Overworld/Nether&End) loaded live
  from repository, local custom_blockset.json as editable working copy
- Add Lab-vs-RGB sanity check with automatic fallback when mismatch detected
- Add luminance-only matching mode for Grayscale blockset
- Rewrite BlockColorRegistry with URL-based loading and Gson parsing
- Rewrite ChunkProcessor with T+- CachedChunkData groundHeight() integration;
  add SatelliteBlockPopulator as primary generation pipeline hook
- Add dual processing path: BlockPopulator (primary) + ChunkLoadEvent (fallback)
- Add hot-switchable tile sources and blocksets via /tsm commands
- Add async version check against GitHub Releases API
- Switch Bing tile URL to {u} dynamic format (TileQuadKey.toBingU)
- Add yandex and tw tile sources to default config
- Update config defaults: max_concurrent_requests=2, cache_size=1000,
  surface_block_mask=false, active_blockset="Default"
- Update Paper API dependency to 1.21.8, bump Lombok to 1.18.42
- Rewrite README: fix config example, flowchart, Python tool path;
  add credits for hueblocks/jar2blockset
- Translate assets2blockset.py comments to English; attribute derivation

+++ src/main/java/com/mndk/tilemapper/config/TileSource.java
@@ -1,3 +1,3 @@
 package com.mndk.tilemapper.config;
 
-public record TileSource(String url, int zoom, int offsetX, int offsetZ) {}
+public record TileSource(String url, int zoom, double offsetX, double offsetZ) {}
